Gate editor spatial module at the module boundary (per Shane's suggestion).
