+++ libs/elodin-editor/src/spatial.rs
@@ -25,16 +25,12 @@
 
 use bevy::{math::DVec3, prelude::*};
 
-#[cfg(feature = "big_space")]
 pub use big_space::prelude::{BigSpace, CellCoord as GridCell, FloatingOrigin, Grid};
 
-#[cfg(feature = "big_space")]
 pub use bevy_geo_frames::big_space::apply_big_translation;
 
-#[cfg(feature = "big_space")]
 pub use big_space::grid::propagation::LowPrecisionRoot;
 
-#[cfg(feature = "big_space")]
 pub mod debug {
     use bevy::prelude::*;
 
@@ -47,13 +43,11 @@ pub mod debug {
     }
 }
 
-#[cfg(feature = "big_space")]
 #[derive(Resource, Clone, Debug)]
 pub struct FloatingOriginSettings {
     grid: Grid,
 }
 
-#[cfg(feature = "big_space")]
 impl FloatingOriginSettings {
     pub fn new(grid_edge_length: f32, switching_threshold: f32) -> Self {
         Self {
@@ -81,19 +75,15 @@ impl FloatingOriginSettings {
 /// Marker for the unique `BigSpace` root entity spawned by
 /// [`FloatingOriginPlugin`]. Lets [`attach_parentless_grid_cells`] adopt
 /// any new grid-bearing entity without having to look up the root by name.
-#[cfg(feature = "big_space")]
 #[derive(Component)]
 pub struct BigSpaceRoot;
 
-#[cfg(feature = "big_space")]
 type ParentlessGridCellFilter = (With<GridCell>, Without<ChildOf>, Without<BigSpaceRoot>);
 
-#[cfg(feature = "big_space")]
 pub struct FloatingOriginPlugin {
     settings: FloatingOriginSettings,
 }
 
-#[cfg(feature = "big_space")]
 impl FloatingOriginPlugin {
     pub fn new(grid_edge_length: f64, switching_threshold: f64) -> Self {
         Self {
@@ -105,7 +95,6 @@ impl FloatingOriginPlugin {
     }
 }
 
-#[cfg(feature = "big_space")]
 impl Plugin for FloatingOriginPlugin {
     fn build(&self, app: &mut App) {
         app.insert_resource(self.settings.clone())
@@ -115,7 +104,6 @@ impl Plugin for FloatingOriginPlugin {
     }
 }
 
-#[cfg(feature = "big_space")]
 pub fn setup_floating_origin(mut commands: Commands, settings: Res<FloatingOriginSettings>) {
     let root = commands
         .spawn((
@@ -140,7 +128,6 @@ pub fn setup_floating_origin(mut commands: Commands, settings: Res<FloatingOrigi
 /// free from having to know about the root entity when they spawn
 /// grid-aware entities. Runs every frame so newly spawned entities are
 /// re-parented as soon as they are visible to the query.
-#[cfg(feature = "big_space")]
 fn attach_parentless_grid_cells(
     mut commands: Commands,
     roots: Query<Entity, With<BigSpaceRoot>>,
@@ -152,114 +139,5 @@ fn attach_parentless_grid_cells(
 
     for entity in &entities {
         commands.entity(entity).insert(ChildOf(root));
-    }
-}
-
-// -----------------------------------------------------------------------------
-// Fallback path: `not(feature = "big_space")`
-//
-// Provides type-compatible stand-ins so the editor still compiles when
-// big_space is disabled. All grid-cell coordinates collapse to (0,0,0)
-// and no precision is gained at large distances. Dev-only.
-// -----------------------------------------------------------------------------
-
-#[cfg(not(feature = "big_space"))]
-#[derive(Component, Default, Clone, Copy, Debug, PartialEq, Eq)]
-pub struct FloatingOrigin;
-
-#[cfg(not(feature = "big_space"))]
-#[derive(Component, Default, Clone, Copy, Debug, PartialEq, Eq)]
-pub struct GridCell {
-    pub x: i128,
-    pub y: i128,
-    pub z: i128,
-}
-
-#[cfg(not(feature = "big_space"))]
-#[derive(Resource, Clone, Copy, Debug)]
-pub struct FloatingOriginSettings {
-    grid_edge_length: f32,
-}
-
-#[cfg(not(feature = "big_space"))]
-impl FloatingOriginSettings {
-    pub fn translation_to_grid(&self, translation: DVec3) -> (GridCell, Vec3) {
-        (GridCell::default(), translation.as_vec3())
-    }
-
-    pub fn grid_edge_length(&self) -> f32 {
-        self.grid_edge_length
-    }
-
-    pub fn grid_position_double(&self, _grid_cell: &GridCell, transform: &Transform) -> DVec3 {
-        transform.translation.as_dvec3()
-    }
-}
-
-#[cfg(not(feature = "big_space"))]
-pub struct FloatingOriginPlugin {
-    grid_edge_length: f32,
-}
-
-#[cfg(not(feature = "big_space"))]
-impl FloatingOriginPlugin {
-    pub fn new(grid_edge_length: f64, _switch_distance: f64) -> Self {
-        Self {
-            grid_edge_length: grid_edge_length as f32,
-        }
-    }
-}
-
-#[cfg(not(feature = "big_space"))]
-impl Plugin for FloatingOriginPlugin {
-    fn build(&self, app: &mut App) {
-        app.insert_resource(FloatingOriginSettings {
-            grid_edge_length: self.grid_edge_length,
-        })
-        .add_systems(Startup, setup_floating_origin);
-    }
-}
-
-#[cfg(not(feature = "big_space"))]
-pub mod debug {
-    use bevy::prelude::*;
-
-    #[derive(Default)]
-    #[allow(dead_code)]
-    pub struct FloatingOriginDebugPlugin;
-
-    impl Plugin for FloatingOriginDebugPlugin {
-        fn build(&self, _app: &mut App) {}
-    }
-}
-
-#[cfg(not(feature = "big_space"))]
-#[derive(Component, Default, Clone, Copy, Debug)]
-pub struct LowPrecisionRoot;
-
-#[cfg(not(feature = "big_space"))]
-pub fn setup_floating_origin(mut commands: Commands) {
-    commands.spawn((
-        FloatingOrigin,
-        GridCell::default(),
-        Transform::IDENTITY,
-        Name::new("floating origin"),
-    ));
-}
-
-#[cfg(not(feature = "big_space"))]
-pub fn apply_big_translation(
-    ctx: ResMut<bevy_geo_frames::GeoContext>,
-    mut q: Query<
-        (&bevy_geo_frames::GeoPosition, &mut Transform, &mut GridCell),
-        Changed<bevy_geo_frames::GeoPosition>,
-    >,
-    floating_origin: Res<FloatingOriginSettings>,
-) {
-    for (geo, mut transform, mut grid_cell) in &mut q {
-        let pos = geo.to_bevy(&ctx);
-        let (new_grid_cell, translation) = floating_origin.translation_to_grid(pos);
-        *grid_cell = new_grid_cell;
-        transform.translation = translation;
     }
 }
